requirements.txt
asgiref==3.8.1
Django==4.2
psycopg2-binary==2.9.10
sqlparse==0.5.3

djangorestframework==3.14
stripe==7.0
twilio==8.10
django-allauth==0.57
django-otp==1.2
django-ratelimit==4.0
python-decouple==3.8
djangorestframework-simplejwt==5.3
celery==5.3
redis==5.0